third_party/blink/web_tests/external/wpt/xhr/formdata/append.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] testFormDataAppendEmptyBlob
  assert_greater_than_equal: expected a "object" but got a "number"
Harness: the test ran to completion.
